corediscount and CA Tax
Every client now has a corediscount. This is automatically filled in for every core line cigar in the order form. If changed it will update in the db.
CA Tax can now be updated from the admin portal. It modifies a file in the server, so it will survive restarts.