Fixed weird bug
Somehow, when I assigned the cc list in emailHandler.js:51, then modified cc, that changed the file itself.
Well, the thing is it didn't change the file, but somehow it changed what the api call GET orders/emails returns. I have no idea how. But, adding .concat() to ensure a copy of the array is being made fixed it, so... c'est la vie.